pyproject.toml
[tool.poetry]
name = "formerbox"
version = "0.1.13"
description = ""
authors = ["mozharovsky <mozharovsky@live.com>"]

[tool.poetry.scripts]
formerbox-cli = 'formerbox.__main__:run_main'

[tool.poetry.dependencies]
# Deep learning toolkit
python = "^3.6.1"
torch = "^1.6.0"
pytorch-lightning = "^1.0.2"

# NLP toolkit packages
transformers = "^3.4.0"
datasets = "^1.1.2"
spacy = "^2.2.4"
sacrebleu = "^1.4.13"

# Machine learning toolkit
numpy = "^1.19.2"
pandas = "^1.0.4"
scikit-learn = "^0.23.1"
scipy = "^1.4.1"

# Misc toolkit packages
tqdm = "^4.46.1"
matplotlib = "^3.2.1"
wandb = "^0.10.5"
boto3 = "^1.15.11"

# System programming toolkit
pydriller = "^1.15.2"
tree_sitter = "^0.1.1"

# Python types magic and extensions
typing_extensions = "^3.7.4"
typing_inspect = "^0.6.0"
more-itertools = "^8.5.0"
jsonlines = "^1.2.0"
overrides = "^3.1.0"
traitlets = "^4.3.3"
PyYAML = "^5.3.1"

# Conditionally installed for backward compatibility with older Python versions
dataclasses = { version = "^0.7", python = "^3.6,<3.7" }

# Optional dependencies
rstcheck = { version = "^3.3.1", optional = true }
docutils = { version = "^0.16", optional = true }
sphinx = { version = "^3.2.1", optional = true }
sphinx-autobuild = { version = "^2020.9.1", optional = true }

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
black = "^20.8b1"
jupyter = "^1.0.0"
jupyterlab = "^2.1.4"
commitizen = "^2.4.0"
cookiecutter = "^1.7.2"
pytest = "^6.1.1"
pytest-ordering = "^0.6"

[tool.commitizen]
name = "cz_formerbox"
version = "0.1.0"
tag_format = "$version"

[tool.pytest.ini_options]
log_cli = 1
log_cli_level = "INFO"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
